+++ pmd-plsql/src/main/java/net/sourceforge/pmd/lang/plsql/rule/design/ExcessivePackageBodyLengthRule.java
@@ -9,7 +9,11 @@ import net.sourceforge.pmd.lang.plsql.ast.ASTPackageBody;
 /**
  * This rule detects when a class exceeds a certain threshold. i.e. if a class
  * has more than 1000 lines of code.
+ *
+ * <p>Equivalent XPath: {@code //PackageBody[@EndLine - @BeginLine > 1000]}</p>
+ * @deprecated Since 7.19.0. Use the rule {@link NcssCountRule} instead.
  */
+@Deprecated
 public class ExcessivePackageBodyLengthRule extends AbstractCounterCheckRule.AbstractLineLengthCheckRule<ASTPackageBody> {
     public ExcessivePackageBodyLengthRule() {
         super(ASTPackageBody.class);