+++ pmd-plsql/src/test/java/net/sourceforge/pmd/lang/plsql/rule/design/ExcessivePackageSpecificationLengthTest.java
@@ -6,6 +6,10 @@ package net.sourceforge.pmd.lang.plsql.rule.design;
 
 import net.sourceforge.pmd.test.PmdRuleTst;
 
+/**
+ * @deprecated Since 7.19.0.
+ */
+@Deprecated
 class ExcessivePackageSpecificationLengthTest extends PmdRuleTst {
     // no additional unit tests
 }